+++ RandomJyrest/src/featureselection/CHCGeneticAlgorithm.java
@@ -263,7 +263,7 @@ public class CHCGeneticAlgorithm
 	    {
 	    	Forest forest = new Forest();
 	    	Map<String, double[]> predictions = forest.main(dataset, numberOfTrees, mtry, p, weights, numberOfThreads, true);
-	    	Map<String, Map<String, Integer>> confusionMatrix = PredictionAnalysis.calculateConfusionMatrix(observationClasses, predictions);
+	    	Map<String, Map<String, Double>> confusionMatrix = PredictionAnalysis.calculateConfusionMatrix(observationClasses, predictions);
 
 	    	double individualFitness = PredictionAnalysis.calculateGMean(confusionMatrix, observationClasses);
